Refactor passkey management by removing the addPasskeyMutationOptions function and integrating its logic directly into the AddPasskeyDialog component. This change simplifies the passkey addition process and enhances error handling and success notifications.
